Tweaked the colors, tweaked some state propagation, added some signalr (not in use yet) to help avoid full re-renders.

+++ API/Entities/HighlightSlot.cs
@@ -16,5 +16,5 @@ public struct RgbaColor
     public int R { get; set; }
     public int G { get; set; }
     public int B { get; set; }
-    public int A { get; set; }
+    public float A { get; set; }
 }
